move transform functions to respective namespaces
