+++ code/tests/stf/CMakeLists.txt
@@ -4,7 +4,6 @@ set(TEST_FILES_PREFIX "cpp/stf")
 
 set(TEST_FILES
     ${TEST_FILES_PREFIX}/alg/interpolation_tests.cpp
-    ${TEST_FILES_PREFIX}/alg/transform_tests.cpp
     ${TEST_FILES_PREFIX}/cam/scamera_tests.cpp
     ${TEST_FILES_PREFIX}/gfx/color_tests.cpp
     ${TEST_FILES_PREFIX}/math/aabb2_tests.cpp
@@ -13,6 +12,7 @@ set(TEST_FILES
     ${TEST_FILES_PREFIX}/math/range_tests.cpp
     ${TEST_FILES_PREFIX}/math/segment2_tests.cpp
     ${TEST_FILES_PREFIX}/math/spherical_tests.cpp
+    ${TEST_FILES_PREFIX}/math/transform_tests.cpp
     ${TEST_FILES_PREFIX}/math/vec2_tests.cpp
     ${TEST_FILES_PREFIX}/math/vec3_tests.cpp
     ${TEST_FILES_PREFIX}/math/vec4_tests.cpp
