+++ code/tests/stf/cpp/stf/cam/scamera_tests.cpp
@@ -28,4 +28,18 @@ namespace stf::cam
         for (scaffolding::scamera::direction_vectors<float> const& test : tests) { scaffolding::scamera::verify(test); }
     }
 
+    TEST(scamera, orbit)
+    {
+        float constexpr half_pi = stff::constants::half_pi;
+        float constexpr pi = stff::constants::pi;
+
+        std::vector<scaffolding::scamera::orbit<float>> tests = 
+        {
+            // initial                                          focus           delta phi                   delta theta             expected
+            { stff::scamera(stff::vec3(0, 0, 10), half_pi, pi), stff::vec3(0),  -stff::constants::half_pi,  stff::constants::pi,    stff::scamera(stff::vec3(0, 10, 0), pi + half_pi, half_pi) },
+        };
+
+        for (scaffolding::scamera::orbit<float> const& test : tests) { scaffolding::scamera::verify(test); }
+    }
+
 } // stf::cam
