+++ code/tests/stf/include/private/stf/math/scaffolding/transform.hpp
@@ -3,9 +3,9 @@
 #include <gtest/gtest.h>
 
 #include <stf/types.hpp>
-#include <stf/alg/transform.hpp>
+#include <stf/math/transform.hpp>
 
-namespace stf::alg::scaffolding::transform
+namespace stf::math::scaffolding::transform
 {
 
     template<typename T>
@@ -19,7 +19,7 @@ namespace stf::alg::scaffolding::transform
     template<typename T>
     void verify(rotate2<T> const& test)
     {
-        ASSERT_EQ(test.expected, alg::rotate(test.initial, test.delta_theta)) << "Failed rotate function";
+        ASSERT_EQ(test.expected, math::rotate(test.initial, test.delta_theta)) << "Failed rotate function";
         ASSERT_EQ(test.expected, math::rotate(test.delta_theta) * test.initial) << "Failed rotate matrix";
     }
 
@@ -35,7 +35,7 @@ namespace stf::alg::scaffolding::transform
     template<typename T>
     void verify(orbit2<T> const& test)
     {
-        ASSERT_EQ(test.expected, alg::orbit(test.initial, test.focus, test.delta_theta)) << "Failed orbit function";
+        ASSERT_EQ(test.expected, math::orbit(test.initial, test.focus, test.delta_theta)) << "Failed orbit function";
     }
 
     template<typename T>
@@ -51,7 +51,7 @@ namespace stf::alg::scaffolding::transform
     void verify(rotate3<T> const& test)
     {
         using homogenized_t = math::vec<T, 4>;
-        ASSERT_EQ(test.expected, alg::rotate(test.initial, test.axis, test.delta_theta)) << "Failed rotate function";
+        ASSERT_EQ(test.expected, math::rotate(test.initial, test.axis, test.delta_theta)) << "Failed rotate function";
         ASSERT_EQ(test.expected, (math::rotate(test.axis, test.delta_theta) * homogenized_t(test.initial, 1)).xyz) << "Failed rotate matrix";
     }
 
@@ -70,24 +70,8 @@ namespace stf::alg::scaffolding::transform
     void verify(orbit3<T> const& test)
     {
         using homogenized_t = math::vec<T, 4>;
-        ASSERT_EQ(test.expected, alg::orbit(test.initial, test.focus, test.right, test.delta_phi, test.delta_theta)) << "Failed orbit function";
+        ASSERT_EQ(test.expected, math::orbit(test.initial, test.focus, test.right, test.delta_phi, test.delta_theta)) << "Failed orbit function";
         ASSERT_EQ(test.expected, (math::orbit(test.focus, test.right, test.delta_phi, test.delta_theta) * homogenized_t(test.initial, 1)).xyz) << "Failed orbit matrix";
     }
 
-    template<typename T>
-    struct orbit_scamera
-    {
-        cam::scamera<T> const initial;
-        math::vec3<T> const focus;
-        T const delta_phi;
-        T const delta_theta;
-        cam::scamera<T> const expected;
-    };
-
-    template<typename T>
-    void verify(orbit_scamera<T> const& test)
-    {
-        ASSERT_EQ(test.expected, alg::orbit(test.initial, test.focus, test.delta_phi, test.delta_theta)) << "Failed scamera orbit function";
-    }
-
-} // stf::alg::scaffolding::transform
+} // stf::math::scaffolding::transform
